if not past yesterdays islamic midnight, show yesterday prayer times
This is to handle the de-syncing of Islamic and Gregorian dates. in a useful way.
if we have passed yesterdays midnight (e.g. it's currently 00:01, and islamic midnight is at 00:30), the gregorian date is now +1, but if the prayer times were to update to reflect the day change, that midnight time would now show the next days, and be inaccurate.
Therefore, until we pass either midnight, or last third (user preference), we still display the day before's date, and show a -1 icon next to the date to highlight that fact.